move trimming to hook
